Refactor: refactor code address PR request

+++ app/helpers/sessions_helper.rb
@@ -7,6 +7,8 @@ module SessionsHelper
   end
 
   def forget(user)
+    return unless user
+
     user.forget
     cookies.delete(:user_id)
     cookies.delete(:remember_token)
@@ -35,6 +37,8 @@ module SessionsHelper
   end
 
   def log_out
+    return unless current_user
+
     forget(current_user)
     session.delete(:user_id)
     @current_user = nil
